+++ examples/custom_multistage_punishments/bot.py
@@ -3,6 +3,7 @@ from AntiSpamTrackerSubclass import MyCustomTracker
 from discord.ext import commands
 
 from antispam import AntiSpamHandler, Options
+from antispam.enums import Library
 from examples.jsonLoader import read_json
 
 bot = commands.Bot(command_prefix=".", intents=discord.Intents.all())
@@ -12,7 +13,9 @@ file = read_json("token")
 
 # Generally you only need/want AntiSpamHandler(bot)
 bot.handler = AntiSpamHandler(
-    bot, options=Options(ignore_bots=False, no_punish=True, use_timeouts=False)
+    bot,
+    library=Library.DPY,
+    options=Options(ignore_bots=False, no_punish=True, use_timeouts=False),
 )
 bot.tracker = MyCustomTracker(bot.handler, 3)
 bot.handler.register_plugin(bot.tracker)